Git status process should die when statusline dies
On large git repos, such as `torvalds/linux` git status takes a lot of time, even if nothing was
changed. On my machine it takes 173 ms, which is more than 98% of time. Therefore it is possible
to enter some simple commands (with shorthand aliases) which may use index while statusline's
`git status` has not exited.

This should have been an impossible bug --- because I thought every children dies when its parent
die, but they were not. After bash acceptes command it prints `PS0` which is configured to trigger
statusline with a deadly signal (SIGIO on configured file descriptor). As majority of time spent
in extending statusline from "fast info" to "all info" is in `git status`, it is likely to stay
alive even after statusline's death --- which could happen only if it was killed directly (then
`git status` will continue to live if not configured properly), or if it committed a suicide after
receiving SIGIO.

`git status` uses special file `index.lock` to lock index and prohibit other git processes to work
with it. As an example, one cannot use `git add` while there is a `git status` on the same repo
somwhere else.

Run many times `git add aaa` after these lines (you can use ArrowUp, Enter for this) to reproduce
the bug:

    $ git clone https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
    Cloning into 'linux'...
    ...
    $ cd linux
    $ touch aaa
    $ git add aaa